refactor: decompose FamilyManager large files into partial classes
Extract partial classes for MainViewModel, CategoryTreeEditor, ImportService,
and PropertiesViewModel. Remove dead sync APIs from Migrator. Unify Export
methods. Fix net48 Split() compatibility.
