+++ src/SmartCon.FamilyManager/Services/LocalCatalog/LocalFamilyImportService.cs
@@ -7,7 +7,7 @@ using SmartCon.Core.Services.Interfaces;
 
 namespace SmartCon.FamilyManager.Services.LocalCatalog;
 
-internal sealed class LocalFamilyImportService : IFamilyImportService
+internal sealed partial class LocalFamilyImportService : IFamilyImportService
 {
     private readonly LocalCatalogDatabase _database;
     private readonly LocalCatalogMigrator _migrator;
@@ -83,20 +83,8 @@ internal sealed class LocalFamilyImportService : IFamilyImportService
             ? await GetNextVersionLabelAsync(existingItem.Id, ct)
             : "v1";
 
-        string relativePath;
-        try
-        {
-            _pathResolver.EnsureFamilyDirectories(catalogItemId, versionLabel, revitVersion);
-            var fileName = metadata.FileName;
-            var absolutePath = _pathResolver.GetRfaFilePath(catalogItemId, versionLabel, revitVersion, fileName);
-            File.Copy(filePath, absolutePath, overwrite: true);
-            File.SetAttributes(absolutePath, File.GetAttributes(absolutePath) | FileAttributes.ReadOnly);
-            relativePath = _pathResolver.GetRelativePath(absolutePath);
-            SmartConLogger.Info($"[Import] Copied to managed storage (read-only): {absolutePath}");
-        }
-        catch (Exception ex)
-        {
-            SmartConLogger.Info($"[Import] Copy FAILED: {ex.Message}");
+        var copyResult = await CopyToManagedStorageAsync(filePath, catalogItemId, versionLabel, revitVersion, metadata);
+        if (!copyResult.Success)
             return new FamilyImportResult(
                 Success: false,
                 CatalogItemId: null,
@@ -104,8 +92,7 @@ internal sealed class LocalFamilyImportService : IFamilyImportService
                 FileId: null,
                 FileName: metadata.FileName,
                 VersionLabel: null,
-                ErrorMessage: $"Failed to copy file to managed storage: {ex.Message}");
-        }
+                ErrorMessage: copyResult.ErrorMessage);
 
         try
         {
@@ -121,7 +108,7 @@ internal sealed class LocalFamilyImportService : IFamilyImportService
 
             try
             {
-                await InsertFileRecordAsync(connection, fileRecordId, relativePath, metadata, revitVersion, now, ct);
+                await InsertFileRecordAsync(connection, fileRecordId, copyResult.RelativePath!, metadata, revitVersion, now, ct);
 
                 if (existingItem is null)
                 {
@@ -164,16 +151,7 @@ internal sealed class LocalFamilyImportService : IFamilyImportService
         }
         catch
         {
-            try
-            {
-                var absPath = Path.Combine(_database.GetDatabaseRoot(), relativePath);
-                if (File.Exists(absPath))
-                    File.Delete(absPath);
-            }
-            catch
-            {
-            }
-
+            await CleanupFileAsync(copyResult.RelativePath);
             throw;
         }
     }
@@ -258,22 +236,6 @@ internal sealed class LocalFamilyImportService : IFamilyImportService
             ErrorCount: errorCount);
     }
 
-    private async Task<FamilyCatalogItem?> FindByNameAsync(string name, CancellationToken ct)
-    {
-        var normalizedName = FamilyNameNormalizer.Normalize(name);
-        using var connection = _database.CreateConnection();
-        await connection.OpenAsync(ct);
-        using var cmd = connection.CreateCommand();
-        cmd.CommandText = "SELECT * FROM catalog_items WHERE normalized_name = @name LIMIT 1";
-        cmd.Parameters.Add(new SqliteParameter("@name", normalizedName));
-
-        using var reader = await cmd.ExecuteReaderAsync(ct);
-        if (!await reader.ReadAsync(ct))
-            return null;
-
-        return LocalCatalogProvider.ReadCatalogItem(reader) with { Tags = [] };
-    }
-
     public async Task<FamilyImportResult> UpdateFamilyAsync(FamilyUpdateRequest request, CancellationToken ct = default)
     {
         await _migrator.MigrateAsync(ct);
@@ -318,19 +280,8 @@ internal sealed class LocalFamilyImportService : IFamilyImportService
         var fileRecordId = Guid.NewGuid().ToString();
         var versionId = Guid.NewGuid().ToString();
 
-        string relativePath;
-        try
-        {
-            _pathResolver.EnsureFamilyDirectories(request.CatalogItemId, versionLabel, revitVersion);
-            var absolutePath = _pathResolver.GetRfaFilePath(request.CatalogItemId, versionLabel, revitVersion, metadata.FileName);
-            File.Copy(filePath, absolutePath, overwrite: true);
-            File.SetAttributes(absolutePath, File.GetAttributes(absolutePath) | FileAttributes.ReadOnly);
-            relativePath = _pathResolver.GetRelativePath(absolutePath);
-            SmartConLogger.Info($"[Update] Copied to managed storage (read-only): {absolutePath}");
-        }
-        catch (Exception ex)
-        {
-            SmartConLogger.Info($"[Update] Copy FAILED: {ex.Message}");
+        var copyResult = await CopyToManagedStorageAsync(filePath, request.CatalogItemId, versionLabel, revitVersion, metadata);
+        if (!copyResult.Success)
             return new FamilyImportResult(
                 Success: false,
                 CatalogItemId: null,
@@ -338,8 +289,7 @@ internal sealed class LocalFamilyImportService : IFamilyImportService
                 FileId: null,
                 FileName: metadata.FileName,
                 VersionLabel: null,
-                ErrorMessage: $"Failed to copy file to managed storage: {ex.Message}");
-        }
+                ErrorMessage: copyResult.ErrorMessage);
 
         try
         {
@@ -355,7 +305,7 @@ internal sealed class LocalFamilyImportService : IFamilyImportService
 
             try
             {
-                await InsertFileRecordAsync(connection, fileRecordId, relativePath, metadata, revitVersion, now, ct);
+                await InsertFileRecordAsync(connection, fileRecordId, copyResult.RelativePath!, metadata, revitVersion, now, ct);
                 await UpdateCatalogItemWithNameAsync(connection, request.CatalogItemId, newName, normalizedName, versionLabel, now, ct);
                 await InsertVersionAsync(connection, versionId, request.CatalogItemId, fileRecordId, versionLabel, metadata, revitVersion, now, ct);
 
@@ -381,180 +331,44 @@ internal sealed class LocalFamilyImportService : IFamilyImportService
         }
         catch
         {
-            try
-            {
-                var absPath = Path.Combine(_database.GetDatabaseRoot(), relativePath);
-                if (File.Exists(absPath))
-                    File.Delete(absPath);
-            }
-            catch
-            {
-            }
-
+            await CleanupFileAsync(copyResult.RelativePath);
             throw;
         }
     }
 
-    private static async Task UpdateCatalogItemWithNameAsync(SqliteConnection connection, string id,
-        string newName, string normalizedName, string versionLabel, DateTimeOffset now, CancellationToken ct)
+    private async Task<CopyResult> CopyToManagedStorageAsync(string sourcePath, string catalogItemId, string versionLabel, int revitVersion, FamilyMetadataExtractionResult metadata)
     {
-        using var cmd = connection.CreateCommand();
-        cmd.CommandText = """
-            UPDATE catalog_items
-            SET name = @name, normalized_name = @normalizedName, current_version_label = @versionLabel, updated_at_utc = @updatedAtUtc
-            WHERE id = @id
-            """;
-        cmd.Parameters.Add(new SqliteParameter("@id", id));
-        cmd.Parameters.Add(new SqliteParameter("@name", newName));
-        cmd.Parameters.Add(new SqliteParameter("@normalizedName", normalizedName));
-        cmd.Parameters.Add(new SqliteParameter("@versionLabel", versionLabel));
-        cmd.Parameters.Add(new SqliteParameter("@updatedAtUtc", now.ToString("o")));
-        await cmd.ExecuteNonQueryAsync(ct);
-    }
-
-    private async Task<FamilyCatalogVersion?> FindVersionByHashAndRevitAsync(string catalogItemId, string sha256, int revitVersion, CancellationToken ct)
-    {
-        using var connection = _database.CreateConnection();
-        await connection.OpenAsync(ct);
-        using var cmd = connection.CreateCommand();
-        cmd.CommandText = """
-            SELECT cv.* FROM catalog_versions cv
-            INNER JOIN family_files ff ON ff.id = cv.file_id
-            WHERE cv.catalog_item_id = @itemId AND ff.sha256 = @sha256 AND cv.revit_major_version = @revitVersion
-            LIMIT 1
-            """;
-        cmd.Parameters.Add(new SqliteParameter("@itemId", catalogItemId));
-        cmd.Parameters.Add(new SqliteParameter("@sha256", sha256));
-        cmd.Parameters.Add(new SqliteParameter("@revitVersion", revitVersion));
-
-        using var reader = await cmd.ExecuteReaderAsync(ct);
-        if (!await reader.ReadAsync(ct))
-            return null;
-
-        return new FamilyCatalogVersion(
-            Id: reader.GetString(reader.GetOrdinal("id")),
-            CatalogItemId: reader.GetString(reader.GetOrdinal("catalog_item_id")),
-            FileId: reader.GetString(reader.GetOrdinal("file_id")),
-            VersionLabel: reader.GetString(reader.GetOrdinal("version_label")),
-            Sha256: reader.GetString(reader.GetOrdinal("sha256")),
-            RevitMajorVersion: reader.GetInt32(reader.GetOrdinal("revit_major_version")),
-            TypesCount: reader.IsDBNull(reader.GetOrdinal("types_count"))
-                ? null
-                : reader.GetInt32(reader.GetOrdinal("types_count")),
-            ParametersCount: reader.IsDBNull(reader.GetOrdinal("parameters_count"))
-                ? null
-                : reader.GetInt32(reader.GetOrdinal("parameters_count")),
-            PublishedAtUtc: DateTimeOffset.Parse(reader.GetString(reader.GetOrdinal("published_at_utc"))));
-    }
-
-    private async Task<string> GetNextVersionLabelAsync(string catalogItemId, CancellationToken ct)
-    {
-        using var connection = _database.CreateConnection();
-        await connection.OpenAsync(ct);
-        using var cmd = connection.CreateCommand();
-        cmd.CommandText = "SELECT version_label FROM catalog_versions WHERE catalog_item_id = @itemId ORDER BY published_at_utc DESC LIMIT 1";
-        cmd.Parameters.Add(new SqliteParameter("@itemId", catalogItemId));
-
-        var result = await cmd.ExecuteScalarAsync(ct);
-        if (result is string label && label.StartsWith("v") && int.TryParse(label[1..], out var num))
+        try
         {
-            return $"v{num + 1}";
+            _pathResolver.EnsureFamilyDirectories(catalogItemId, versionLabel, revitVersion);
+            var absolutePath = _pathResolver.GetRfaFilePath(catalogItemId, versionLabel, revitVersion, metadata.FileName);
+            File.Copy(sourcePath, absolutePath, overwrite: true);
+            File.SetAttributes(absolutePath, File.GetAttributes(absolutePath) | FileAttributes.ReadOnly);
+            var relativePath = _pathResolver.GetRelativePath(absolutePath);
+            SmartConLogger.Info($"[Import] Copied to managed storage (read-only): {absolutePath}");
+            return new CopyResult(true, relativePath, null);
         }
-
-        return "v2";
+        catch (Exception ex)
+        {
+            SmartConLogger.Info($"[Import] Copy FAILED: {ex.Message}");
+            return new CopyResult(false, null, $"Failed to copy file to managed storage: {ex.Message}");
+        }
     }
 
-    private static async Task InsertFileRecordAsync(SqliteConnection connection, string id,
-        string relativePath, FamilyMetadataExtractionResult metadata,
-        int revitVersion, DateTimeOffset now, CancellationToken ct)
+    private async Task CleanupFileAsync(string? relativePath)
     {
-        using var cmd = connection.CreateCommand();
-        cmd.CommandText = """
-            INSERT INTO family_files (id, relative_path, file_name, size_bytes, sha256, revit_major_version, imported_at_utc)
-            VALUES (@id, @relativePath, @fileName, @sizeBytes, @sha256, @revitVersion, @importedAtUtc)
-            """;
-        cmd.Parameters.Add(new SqliteParameter("@id", id));
-        cmd.Parameters.Add(new SqliteParameter("@relativePath", relativePath));
-        cmd.Parameters.Add(new SqliteParameter("@fileName", metadata.FileName));
-        cmd.Parameters.Add(new SqliteParameter("@sizeBytes", metadata.FileSizeBytes));
-        cmd.Parameters.Add(new SqliteParameter("@sha256", metadata.Sha256));
-        cmd.Parameters.Add(new SqliteParameter("@revitVersion", revitVersion));
-        cmd.Parameters.Add(new SqliteParameter("@importedAtUtc", now.ToString("o")));
-        await cmd.ExecuteNonQueryAsync(ct);
+        if (string.IsNullOrEmpty(relativePath)) return;
+        try
+        {
+            var absPath = Path.Combine(_database.GetDatabaseRoot(), relativePath);
+            if (File.Exists(absPath))
+                File.Delete(absPath);
+        }
+        catch
+        {
+            // ignored
+        }
     }
 
-    private static async Task InsertCatalogItemAsync(SqliteConnection connection, string id,
-        string normalizedName, FamilyImportRequest request, DateTimeOffset now,
-        string versionLabel, CancellationToken ct)
-    {
-        using var cmd = connection.CreateCommand();
-        cmd.CommandText = """
-            INSERT INTO catalog_items (id, name, normalized_name, description, category_name, category_id, manufacturer, content_status, current_version_label, published_by, created_at_utc, updated_at_utc)
-            VALUES (@id, @name, @normalizedName, @description, @categoryName, @categoryId, @manufacturer, @status, @versionLabel, @publishedBy, @createdAtUtc, @updatedAtUtc)
-            """;
-        cmd.Parameters.Add(new SqliteParameter("@id", id));
-        cmd.Parameters.Add(new SqliteParameter("@name", Path.GetFileNameWithoutExtension(request.FilePath)));
-        cmd.Parameters.Add(new SqliteParameter("@normalizedName", normalizedName));
-        cmd.Parameters.Add(new SqliteParameter("@description", request.Description ?? (object)DBNull.Value));
-        cmd.Parameters.Add(new SqliteParameter("@categoryName", request.Category ?? (object)DBNull.Value));
-        cmd.Parameters.Add(new SqliteParameter("@categoryId", request.CategoryId ?? (object)DBNull.Value));
-        cmd.Parameters.Add(new SqliteParameter("@manufacturer", DBNull.Value));
-        cmd.Parameters.Add(new SqliteParameter("@status", ContentStatus.Active.ToString()));
-        cmd.Parameters.Add(new SqliteParameter("@versionLabel", versionLabel));
-        cmd.Parameters.Add(new SqliteParameter("@publishedBy", DBNull.Value));
-        cmd.Parameters.Add(new SqliteParameter("@createdAtUtc", now.ToString("o")));
-        cmd.Parameters.Add(new SqliteParameter("@updatedAtUtc", now.ToString("o")));
-        await cmd.ExecuteNonQueryAsync(ct);
-    }
-
-    private static async Task UpdateCatalogItemVersionAsync(SqliteConnection connection, string id,
-        string versionLabel, DateTimeOffset now, CancellationToken ct)
-    {
-        using var cmd = connection.CreateCommand();
-        cmd.CommandText = """
-            UPDATE catalog_items SET current_version_label = @versionLabel, updated_at_utc = @updatedAtUtc WHERE id = @id
-            """;
-        cmd.Parameters.Add(new SqliteParameter("@id", id));
-        cmd.Parameters.Add(new SqliteParameter("@versionLabel", versionLabel));
-        cmd.Parameters.Add(new SqliteParameter("@updatedAtUtc", now.ToString("o")));
-        await cmd.ExecuteNonQueryAsync(ct);
-    }
-
-    private static async Task InsertVersionAsync(SqliteConnection connection, string versionId,
-        string catalogItemId, string fileId, string versionLabel,
-        FamilyMetadataExtractionResult metadata, int revitVersion,
-        DateTimeOffset now, CancellationToken ct)
-    {
-        using var cmd = connection.CreateCommand();
-        cmd.CommandText = """
-            INSERT INTO catalog_versions (id, catalog_item_id, file_id, version_label, sha256, revit_major_version, types_count, parameters_count, published_at_utc)
-            VALUES (@id, @catalogItemId, @fileId, @versionLabel, @sha256, @revitMajorVersion, @typesCount, @parametersCount, @publishedAtUtc)
-            """;
-        cmd.Parameters.Add(new SqliteParameter("@id", versionId));
-        cmd.Parameters.Add(new SqliteParameter("@catalogItemId", catalogItemId));
-        cmd.Parameters.Add(new SqliteParameter("@fileId", fileId));
-        cmd.Parameters.Add(new SqliteParameter("@versionLabel", versionLabel));
-        cmd.Parameters.Add(new SqliteParameter("@sha256", metadata.Sha256));
-        cmd.Parameters.Add(new SqliteParameter("@revitMajorVersion", revitVersion));
-        cmd.Parameters.Add(new SqliteParameter("@typesCount",
-            metadata.Types is not null ? (object)metadata.Types.Count : DBNull.Value));
-        cmd.Parameters.Add(new SqliteParameter("@parametersCount",
-            metadata.Parameters is not null ? (object)metadata.Parameters.Count : DBNull.Value));
-        cmd.Parameters.Add(new SqliteParameter("@publishedAtUtc", now.ToString("o")));
-        await cmd.ExecuteNonQueryAsync(ct);
-    }
-
-    private static async Task InsertTagAsync(SqliteConnection connection, string catalogItemId, string tag, CancellationToken ct)
-    {
-        var normalizedTag = FamilySearchNormalizer.Normalize(tag);
-        using var cmd = connection.CreateCommand();
-        cmd.CommandText = """
-            INSERT OR IGNORE INTO catalog_tags (catalog_item_id, tag, normalized_tag)
-            VALUES (@catalogItemId, @tag, @normalizedTag)
-            """;
-        cmd.Parameters.Add(new SqliteParameter("@catalogItemId", catalogItemId));
-        cmd.Parameters.Add(new SqliteParameter("@tag", tag));
-        cmd.Parameters.Add(new SqliteParameter("@normalizedTag", normalizedTag));
-        await cmd.ExecuteNonQueryAsync(ct);
-    }
+    private readonly record struct CopyResult(bool Success, string? RelativePath, string? ErrorMessage);
 }
