+++ src/SmartCon.FamilyManager/ViewModels/CategoryTreeEditorViewModel.cs
@@ -1,7 +1,5 @@
 using System.Collections.ObjectModel;
-using System.IO;
 using System.Linq;
-using System.Text.Json;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using SmartCon.Core.Logging;
@@ -223,186 +221,6 @@ public sealed partial class CategoryTreeEditorViewModel : ObservableObject, IObs
         }
     }
 
-    private async Task<List<EffectiveCategoryAttribute>> GetDraftEffectiveAttributesAsync(
-        CategoryNodeViewModel node, Dictionary<string, string> categoryNameById)
-    {
-        var result = new Dictionary<string, EffectiveCategoryAttribute>();
-
-        // Step 1: Load base effective attrs
-        if (!node.IsNew)
-        {
-            var dbEffective = await _bindingService.GetEffectiveAttributesAsync(node.CategoryId);
-            foreach (var attr in dbEffective)
-            {
-                result[attr.AttributeId] = attr;
-            }
-
-            // Apply ALL ancestor draft changes recursively
-            await ApplyAncestorDraftChangesAsync(node.ParentId, result);
-        }
-        else if (node.ParentId is not null)
-        {
-            // Draft node: inherit from parent recursively with draft awareness
-            var parent = FindNodeById(RootNodes, node.ParentId);
-            if (parent is not null)
-            {
-                var parentEffective = await GetDraftEffectiveAttributesAsync(parent, categoryNameById);
-                foreach (var attr in parentEffective)
-                {
-                    result[attr.AttributeId] = attr with { IsInherited = true };
-                }
-            }
-            else
-            {
-                var dbEffective = await _bindingService.GetEffectiveAttributesAsync(node.ParentId);
-                foreach (var attr in dbEffective)
-                {
-                    result[attr.AttributeId] = attr with { IsInherited = true };
-                }
-            }
-        }
-
-        // Step 2: Apply own draft changes (add/remove bindings)
-        var ownChanges = _bindingChanges
-            .Where(c =>
-            {
-                var parts = c.Key.Split(new[] { ':' }, 2);
-                return parts.Length == 2 && parts[0] == node.CategoryId;
-            })
-            .ToList();
-
-        foreach (var change in ownChanges)
-        {
-            var parts = change.Key.Split(new[] { ':' }, 2);
-            var attrId = parts[1];
-
-            if (change.Value)
-            {
-                // Added binding
-                if (!result.ContainsKey(attrId))
-                {
-                    var attrDef = await _attributeDefRepository.GetByIdAsync(attrId);
-                    if (attrDef is not null)
-                    {
-                        result[attrId] = new EffectiveCategoryAttribute(
-                            attrId,
-                            attrDef.Name,
-                            attrDef.Group,
-                            result.Count,
-                            true,
-                            false, // direct binding, not inherited
-                            node.CategoryId);
-                    }
-                }
-            }
-            else
-            {
-                // Removed binding
-                result.Remove(attrId);
-            }
-        }
-
-        return [..result.Values];
-    }
-
-    private List<CategoryAttributeBinding> GetDraftDirectBindings(CategoryNodeViewModel node)
-    {
-        var result = new List<CategoryAttributeBinding>();
-
-        foreach (var change in _bindingChanges)
-        {
-            var parts = change.Key.Split(new[] { ':' }, 2);
-            if (parts.Length != 2 || parts[0] != node.CategoryId) continue;
-
-            if (change.Value)
-            {
-                result.Add(new CategoryAttributeBinding(
-                    Guid.NewGuid().ToString(),
-                    node.CategoryId,
-                    parts[1],
-                    result.Count,
-                    true));
-            }
-        }
-
-        return result;
-    }
-
-    private async Task ApplyAncestorDraftChangesAsync(
-        string? parentId,
-        Dictionary<string, EffectiveCategoryAttribute> result)
-    {
-        if (parentId is null) return;
-
-        // Check if this ancestor has draft changes
-        var parentChanges = _bindingChanges
-            .Where(c => c.Key.StartsWith($"{parentId}:"))
-            .ToList();
-
-        foreach (var change in parentChanges)
-        {
-            var parts = change.Key.Split(new[] { ':' }, 2);
-            if (parts.Length != 2) continue;
-            var attrId = parts[1];
-            var isBound = change.Value;
-
-            if (isBound && !result.ContainsKey(attrId))
-            {
-                var attrDef = await _attributeDefRepository.GetByIdAsync(attrId);
-                if (attrDef is not null)
-                {
-                    result[attrId] = new EffectiveCategoryAttribute(
-                        attrId,
-                        attrDef.Name,
-                        attrDef.Group,
-                        result.Count,
-                        true,
-                        true,
-                        parentId);
-                }
-            }
-            else if (!isBound && result.ContainsKey(attrId) && result[attrId].SourceCategoryId == parentId)
-            {
-                result.Remove(attrId);
-            }
-        }
-
-        // Recursively apply grandparent's draft changes
-        var parent = FindNodeById(RootNodes, parentId);
-        if (parent is not null)
-        {
-            await ApplyAncestorDraftChangesAsync(parent.ParentId, result);
-        }
-        else
-        {
-            // Parent only in DB - get its parentId from DB
-            var parentNode = await _categoryRepository.GetByIdAsync(parentId);
-            await ApplyAncestorDraftChangesAsync(parentNode?.ParentId, result);
-        }
-    }
-
-    private static CategoryNodeViewModel? FindNodeById(ObservableCollection<CategoryNodeViewModel> nodes, string categoryId)
-    {
-        foreach (var node in nodes)
-        {
-            if (node.CategoryId == categoryId) return node;
-            var found = FindNodeById(node.Children, categoryId);
-            if (found is not null) return found;
-        }
-        return null;
-    }
-
-    private static CategoryNodeViewModel? FindNodeById(ObservableCollection<CatalogTreeNodeViewModel> nodes, string categoryId)
-    {
-        foreach (var node in nodes)
-        {
-            if (node is CategoryNodeViewModel cat && cat.CategoryId == categoryId) return cat;
-            var found = FindNodeById(node.Children, categoryId);
-            if (found is not null) return found;
-        }
-        return null;
-    }
-
     internal void HandleBindingToggle(AttributeListItemViewModel item, bool shouldBeBound)
     {
         if (SelectedNode is null) return;
@@ -456,333 +274,6 @@ public sealed partial class CategoryTreeEditorViewModel : ObservableObject, IObs
 
         if (SelectedNode is not null)
             await LoadAttributesForCategoryAsync(SelectedNode);
-    }
-
-    [RelayCommand]
-    private async Task ExportCategoriesAsync()
-    {
-        var title = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_ExportTree) ?? "Export Categories";
-        var path = _dialogService.ShowSaveJsonDialog(title, "categories");
-        if (path is null) return;
-
-        try
-        {
-            var package = await _packageService.ExportCategoriesAsync();
-            var json = JsonSerializer.Serialize(package, new JsonSerializerOptions { WriteIndented = true });
-            await Task.Run(() => File.WriteAllText(path, json));
-            StatusMessage = string.Format(LanguageManager.GetString(StringLocalization.Keys.FM_CTE_Exported) ?? "Exported to {0}", path);
-        }
-        catch (Exception ex)
-        {
-            StatusMessage = string.Format(LanguageManager.GetString(StringLocalization.Keys.FM_ImportError) ?? "Error: {0}", ex.Message);
-        }
-    }
-
-    [RelayCommand]
-    private async Task ExportAttributesAsync()
-    {
-        var title = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_ExportAttrs) ?? "Export Attributes";
-        var path = _dialogService.ShowSaveJsonDialog(title, "attributes");
-        if (path is null) return;
-
-        try
-        {
-            var package = await _packageService.ExportAttributesAsync();
-            var json = JsonSerializer.Serialize(package, new JsonSerializerOptions { WriteIndented = true });
-            await Task.Run(() => File.WriteAllText(path, json));
-            StatusMessage = string.Format(LanguageManager.GetString(StringLocalization.Keys.FM_CTE_Exported) ?? "Exported to {0}", path);
-        }
-        catch (Exception ex)
-        {
-            StatusMessage = string.Format(LanguageManager.GetString(StringLocalization.Keys.FM_ImportError) ?? "Error: {0}", ex.Message);
-        }
-    }
-
-    [RelayCommand]
-    private async Task ExportFullAsync()
-    {
-        var title = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_ExportFull) ?? "Export Full Package";
-        var path = _dialogService.ShowSaveJsonDialog(title, "smartcon-metadata");
-        if (path is null) return;
-
-        try
-        {
-            var package = await _packageService.ExportFullAsync();
-            var json = JsonSerializer.Serialize(package, new JsonSerializerOptions { WriteIndented = true });
-            await Task.Run(() => File.WriteAllText(path, json));
-            StatusMessage = string.Format(LanguageManager.GetString(StringLocalization.Keys.FM_CTE_Exported) ?? "Exported to {0}", path);
-        }
-        catch (Exception ex)
-        {
-            StatusMessage = string.Format(LanguageManager.GetString(StringLocalization.Keys.FM_ImportError) ?? "Error: {0}", ex.Message);
-        }
-    }
-
-    [RelayCommand]
-    private void ImportFromJson()
-    {
-        var title = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_ImportTree) ?? "Import Metadata";
-        var path = _dialogService.ShowOpenJsonDialog(title);
-        if (path is null) return;
-
-        try
-        {
-            var json = File.ReadAllText(path);
-            var package = JsonSerializer.Deserialize<FamilyMetadataPackage>(json);
-            if (package is null) return;
-
-            var importedNodes = new List<CategoryNodeViewModel>();
-            foreach (var cat in package.Categories)
-            {
-                importedNodes.AddRange(ImportCategoryNode(cat, null, 0));
-            }
-
-            RootNodes = new ObservableCollection<CategoryNodeViewModel>(importedNodes);
-            _bindingChanges.Clear();
-            _pendingImportPackage = package;
-            SelectedNode = null;
-            UpdateHasUnsavedChanges();
-            StatusMessage = string.Format(
-                LanguageManager.GetString(StringLocalization.Keys.FM_CTE_Imported) ?? "Imported {0} categories",
-                importedNodes.Count);
-        }
-        catch (Exception ex)
-        {
-            StatusMessage = string.Format(LanguageManager.GetString(StringLocalization.Keys.FM_ImportError) ?? "Import error: {0}", ex.Message);
-        }
-    }
-
-    private static List<CategoryNodeViewModel> ImportCategoryNode(FamilyMetadataCategoryNode node, string? parentId, int sortOrder)
-    {
-        var result = new List<CategoryNodeViewModel>();
-        var categoryId = Guid.NewGuid().ToString();
-        var vm = new CategoryNodeViewModel(categoryId, node.Name, parentId, node.Name)
-        {
-            SortOrder = sortOrder,
-            OriginalSortOrder = sortOrder,
-            IsNew = true,
-            IsDirty = true
-        };
-        result.Add(vm);
-
-        for (int i = 0; i < node.Children.Count; i++)
-        {
-            var children = ImportCategoryNode(node.Children[i], categoryId, i);
-            foreach (var child in children)
-            {
-                vm.Children.Add(child);
-            }
-        }
-
-        return result;
-    }
-
-    [RelayCommand]
-    private void AddRoot()
-    {
-        var title = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_AddCategory) ?? "Add Category";
-        var prompt = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_CategoryName) ?? "Category name:";
-        var name = _dialogService.ShowInputDialog(title, prompt);
-        if (string.IsNullOrWhiteSpace(name)) return;
-
-        var sortOrder = RootNodes.Count;
-        var categoryId = Guid.NewGuid().ToString();
-        var vm = new CategoryNodeViewModel(categoryId, name!, null, name!)
-        {
-            SortOrder = sortOrder,
-            OriginalSortOrder = sortOrder,
-            IsNew = true,
-            IsDirty = true
-        };
-        RootNodes.Add(vm);
-        SelectedNode = vm;
-        vm.IsSelected = true;
-        UpdateHasUnsavedChanges();
-    }
-
-    [RelayCommand]
-    private void AddChild()
-    {
-        if (SelectedNode is not CategoryNodeViewModel parent) return;
-
-        var title = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_AddSubcategory) ?? "Add Subcategory";
-        var prompt = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_SubcategoryName) ?? "Subcategory name:";
-        var name = _dialogService.ShowInputDialog(title, prompt);
-        if (string.IsNullOrWhiteSpace(name)) return;
-
-        var sortOrder = parent.Children.Count;
-        var categoryId = Guid.NewGuid().ToString();
-        var fullPath = $"{parent.FullPath}/{name}";
-        var childVm = new CategoryNodeViewModel(categoryId, name!, parent.CategoryId, fullPath)
-        {
-            SortOrder = sortOrder,
-            OriginalSortOrder = sortOrder,
-            IsNew = true,
-            IsDirty = true
-        };
-        parent.Children.Add(childVm);
-        parent.IsExpanded = true;
-        SelectedNode = childVm;
-        childVm.IsSelected = true;
-        UpdateHasUnsavedChanges();
-    }
-
-    [RelayCommand]
-    private void Rename()
-    {
-        if (SelectedNode is not CategoryNodeViewModel node) return;
-
-        var title = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_Rename) ?? "Rename Category";
-        var prompt = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_NewName) ?? "New name:";
-        var newName = _dialogService.ShowInputDialog(title, prompt, node.DisplayName);
-        if (string.IsNullOrWhiteSpace(newName) || newName == node.DisplayName) return;
-
-        node.DisplayName = newName!;
-        node.IsDirty = true;
-        UpdateHasUnsavedChanges();
-    }
-
-    [RelayCommand]
-    private void Delete()
-    {
-        if (SelectedNode is not CategoryNodeViewModel node) return;
-
-        var count = node.FamilyCount;
-        var noCat = LanguageManager.GetString(StringLocalization.Keys.FM_NoCategory) ?? "No category";
-        var message = count > 0
-            ? string.Format(LanguageManager.GetString(StringLocalization.Keys.FM_CTE_DeleteConfirm) ?? "Delete \"{0}\"? {1} families will be moved to \"{2}\".", node.DisplayName, count, noCat)
-            : string.Format(LanguageManager.GetString(StringLocalization.Keys.FM_CTE_DeleteEmpty) ?? "Delete \"{0}\"?", node.DisplayName);
-
-        var delTitle = LanguageManager.GetString(StringLocalization.Keys.FM_CTE_DeleteCategory) ?? "Delete Category";
-        if (!_dialogService.ShowConfirmation(delTitle, message)) return;
-
-        node.IsDeleted = true;
-        _pendingCategoryDeletions.Add(node);
-        RemoveNodeFromTree(node.CategoryId);
-        SelectedNode = null;
-        UpdateHasUnsavedChanges();
-    }
-
-    private void RemoveNodeFromTree(string categoryId)
-    {
-        for (var i = 0; i < RootNodes.Count; i++)
-        {
-            if (RootNodes[i].CategoryId == categoryId)
-            {
-                RootNodes.RemoveAt(i);
-                return;
-            }
-            if (TryRemoveNode(RootNodes[i].Children, categoryId))
-                return;
-        }
-    }
-
-    internal static bool TryRemoveNode(ObservableCollection<CatalogTreeNodeViewModel> nodes, string categoryId)
-    {
-        for (var i = 0; i < nodes.Count; i++)
-        {
-            if (nodes[i] is CategoryNodeViewModel cat && cat.CategoryId == categoryId)
-            {
-                nodes.RemoveAt(i);
-                return true;
-            }
-            if (TryRemoveNode(nodes[i].Children, categoryId))
-                return true;
-        }
-        return false;
-    }
-
-    [RelayCommand]
-    private void MoveNode((CategoryNodeViewModel Node, CategoryNodeViewModel? NewParent, int SortOrder) args)
-    {
-        var (node, newParent, sortOrder) = args;
-        node.ParentId = newParent?.CategoryId;
-        node.SortOrder = sortOrder;
-        node.IsDirty = true;
-    }
-
-    internal static List<CategoryTreeImportData.CategoryImportItem> BuildExportTree(CategoryTree tree, string? parentId)
-    {
-        var result = new List<CategoryTreeImportData.CategoryImportItem>();
-        foreach (var node in tree.GetChildren(parentId))
-        {
-            result.Add(new CategoryTreeImportData.CategoryImportItem
-            {
-                Name = node.Name,
-                Children = BuildExportTree(tree, node.Id)
-            });
-        }
-        return result;
-    }
-
-    internal static List<CategoryNode> FlattenImportData(List<CategoryTreeImportData.CategoryImportItem> items, string? parentId, int startOrder)
-    {
-        var result = new List<CategoryNode>();
-        for (var i = 0; i < items.Count; i++)
-        {
-            var item = items[i];
-            var id = Guid.NewGuid().ToString();
-            var path = parentId is null ? item.Name : $"{parentId}/{item.Name}";
-            result.Add(new CategoryNode(id, item.Name, parentId, startOrder + i, path, DateTimeOffset.UtcNow));
-            if (item.Children?.Count > 0)
-            {
-                result.AddRange(FlattenImportData(item.Children, id, 0));
-            }
-        }
-        return result;
-    }
-
-    private string BuildCategoryPath(CategoryNodeViewModel target)
-    {
-        var path = new List<string>();
-        foreach (var root in RootNodes)
-        {
-            path.Add(root.DisplayName);
-            if (root.CategoryId == target.CategoryId) break;
-            if (SearchPathInChildren(root.Children, target.CategoryId, path)) break;
-            path.RemoveAt(path.Count - 1);
-        }
-        return string.Join(" > ", path);
-    }
-
-    private static bool SearchPathInChildren(IList<CatalogTreeNodeViewModel> children, string targetId, List<string> path)
-    {
-        foreach (var child in children)
-        {
-            if (child is not CategoryNodeViewModel cat) continue;
-            path.Add(cat.DisplayName);
-            if (cat.CategoryId == targetId) return true;
-            if (SearchPathInChildren(cat.Children, targetId, path)) return true;
-            path.RemoveAt(path.Count - 1);
-        }
-        return false;
-    }
-
-    [RelayCommand]
-    private void OnSelectedItemChanged(object? selectedItem)
-    {
-        SelectedNode = selectedItem as CategoryNodeViewModel;
-    }
-
-    [RelayCommand]
-    private void ContextMenuRename()
-    {
-        if (SelectedNode is not CategoryNodeViewModel node) return;
-        Rename();
-    }
-
-    [RelayCommand]
-    private void ContextMenuAddChild()
-    {
-        if (SelectedNode is not CategoryNodeViewModel parent) return;
-        AddChild();
-    }
-
-    [RelayCommand]
-    private void ContextMenuDelete()
-    {
-        if (SelectedNode is not CategoryNodeViewModel node) return;
-        Delete();
     }
 
     [RelayCommand]
