+++ src/SmartCon.FamilyManager/ViewModels/FamilyManagerMainViewModel.Database.cs
@@ -1,0 +1,223 @@
+using System.Collections.ObjectModel;
+using CommunityToolkit.Mvvm.Input;
+using SmartCon.Core.Logging;
+using SmartCon.Core.Models.FamilyManager;
+using SmartCon.Core.Services;
+using SmartCon.Core.Services.Interfaces;
+using SmartCon.UI;
+
+namespace SmartCon.FamilyManager.ViewModels;
+
+public sealed partial class FamilyManagerMainViewModel
+{
+    private void RefreshConnections()
+    {
+        _suppressConnectionChanged = true;
+        try
+        {
+            var list = _databaseManager.ListConnections();
+            Connections = new ObservableCollection<DatabaseConnection>(list);
+            var active = _databaseManager.GetActiveConnection();
+            SelectedConnection = Connections.FirstOrDefault(c => c.Id == active?.Id);
+        }
+        finally
+        {
+            _suppressConnectionChanged = false;
+        }
+    }
+
+    private void OnActiveDatabaseChanged(object? sender, string connectionId)
+    {
+        RefreshConnections();
+        _ = LoadTreeAsync();
+    }
+
+    partial void OnSelectedConnectionChanged(DatabaseConnection? value)
+    {
+        if (value is null) return;
+        if (_suppressConnectionChanged) return;
+        var active = _databaseManager.GetActiveConnection();
+        if (active?.Id == value.Id) return;
+        _ = SwitchDatabaseAsync(value.Id);
+    }
+
+    private async Task SwitchDatabaseAsync(string connectionId)
+    {
+        IsLoading = true;
+        try
+        {
+            var success = await _databaseManager.SwitchDatabaseAsync(connectionId);
+            if (success)
+            {
+                var conn = Connections.FirstOrDefault(c => c.Id == connectionId);
+                StatusMessage = string.Format(
+                    LanguageManager.GetString(StringLocalization.Keys.FM_DbSwitched) ?? "Switched to: {0}",
+                    conn?.Name ?? connectionId);
+            }
+            else
+            {
+                StatusMessage = LanguageManager.GetString(StringLocalization.Keys.FM_DbSwitchError) ?? "Error switching database";
+                RefreshConnections();
+            }
+        }
+        catch (Exception ex)
+        {
+            StatusMessage = $"{LanguageManager.GetString(StringLocalization.Keys.FM_DbSwitchError) ?? "Error switching database"}: {ex.Message}";
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+
+    [RelayCommand]
+    private async Task CreateDatabaseAsync()
+    {
+        var path = _dialogService.ShowFolderBrowserDialog(
+            LanguageManager.GetString(StringLocalization.Keys.FM_DbSelectPath) ?? "Select parent folder for database");
+        if (string.IsNullOrWhiteSpace(path)) return;
+
+        var name = _dialogService.ShowInputDialog(
+            LanguageManager.GetString(StringLocalization.Keys.FM_DbNewTitle) ?? "New Database",
+            LanguageManager.GetString(StringLocalization.Keys.FM_DbNewPrompt) ?? "Enter database name:",
+            LanguageManager.GetString(StringLocalization.Keys.FM_DbNewDefault) ?? "New Catalog");
+
+        if (string.IsNullOrWhiteSpace(name)) return;
+
+        IsLoading = true;
+        try
+        {
+            var conn = await _databaseManager.CreateDatabaseAsync(name!.Trim(), path!);
+            RefreshConnections();
+            SelectedConnection = Connections.FirstOrDefault(c => c.Id == conn.Id);
+            StatusMessage = string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_DbCreated) ?? "Database \"{0}\" created at {1}",
+                conn.Name, conn.Path);
+        }
+        catch (Exception ex)
+        {
+            StatusMessage = string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_DbCreateError) ?? "Error creating database: {0}",
+                ex.Message);
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+
+    [RelayCommand]
+    private async Task ConnectDatabaseAsync()
+    {
+        var path = _dialogService.ShowFolderBrowserDialog(
+            LanguageManager.GetString(StringLocalization.Keys.FM_DbSelectPath) ?? "Select database folder");
+        if (string.IsNullOrWhiteSpace(path)) return;
+
+        IsLoading = true;
+        try
+        {
+            var conn = await _databaseManager.ConnectDatabaseAsync(path!);
+            RefreshConnections();
+            SelectedConnection = Connections.FirstOrDefault(c => c.Id == conn.Id);
+            StatusMessage = string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_DbSwitched) ?? "Connected to: {0}",
+                conn.Name);
+        }
+        catch (Exception ex)
+        {
+            StatusMessage = string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_DbCreateError) ?? "Error connecting: {0}",
+                ex.Message);
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+
+    [RelayCommand]
+    private async Task DisconnectDatabaseAsync()
+    {
+        if (SelectedConnection is null) return;
+
+        var connections = _databaseManager.ListConnections();
+        if (connections.Count <= 1)
+        {
+            _dialogService.ShowWarning(
+                LanguageManager.GetString(StringLocalization.Keys.FM_DbDeleteTitle) ?? "Disconnect",
+                LanguageManager.GetString(StringLocalization.Keys.FM_CannotDisconnectOnlyDatabase) ?? "Cannot disconnect the only database.");
+            return;
+        }
+
+        IsLoading = true;
+        try
+        {
+            var success = await _databaseManager.DisconnectDatabaseAsync(SelectedConnection.Id);
+            if (success)
+            {
+                StatusMessage = string.Format(
+                    LanguageManager.GetString(StringLocalization.Keys.FM_DbDeleted) ?? "Disconnected: {0}",
+                    SelectedConnection.Name);
+                RefreshConnections();
+            }
+        }
+        catch (Exception ex)
+        {
+            StatusMessage = string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_ErrorFormat) ?? "Error: {0}",
+                ex.Message);
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+
+    [RelayCommand]
+    private async Task DeleteDatabaseAsync()
+    {
+        if (SelectedConnection is null) return;
+
+        var isActive = _databaseManager.GetActiveConnection()?.Id == SelectedConnection.Id;
+        var connections = _databaseManager.ListConnections();
+        if (isActive && connections.Count <= 1)
+        {
+            _dialogService.ShowWarning(
+                LanguageManager.GetString(StringLocalization.Keys.FM_DbDeleteTitle) ?? "Delete Database",
+                LanguageManager.GetString(StringLocalization.Keys.FM_DbDeleteSingle) ?? "Cannot delete the only database.");
+            return;
+        }
+
+        var confirm = _dialogService.ShowInputDialog(
+            LanguageManager.GetString(StringLocalization.Keys.FM_DbDeleteTitle) ?? "Delete Database",
+            string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_DbDeletePrompt) ?? "Enter \"{0}\" to confirm deletion:",
+                SelectedConnection.Name),
+            "");
+
+        if (confirm != SelectedConnection.Name) return;
+
+        IsLoading = true;
+        try
+        {
+            var success = await _databaseManager.DeleteDatabaseAsync(SelectedConnection.Id);
+            if (success)
+            {
+                StatusMessage = string.Format(
+                    LanguageManager.GetString(StringLocalization.Keys.FM_DbDeleted) ?? "Database \"{0}\" deleted",
+                    SelectedConnection.Name);
+                RefreshConnections();
+            }
+        }
+        catch (Exception ex)
+        {
+            StatusMessage = string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_DbDeleteError) ?? "Error: {0}",
+                ex.Message);
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+}
