+++ src/SmartCon.FamilyManager/ViewModels/FamilyManagerMainViewModel.FamilyEdit.cs
@@ -1,0 +1,238 @@
+using System.Collections.ObjectModel;
+using CommunityToolkit.Mvvm.Input;
+using SmartCon.Core.Logging;
+using SmartCon.Core.Models.FamilyManager;
+using SmartCon.Core.Services;
+using SmartCon.Core.Services.Interfaces;
+using SmartCon.UI;
+using SmartCon.UI.Behaviors;
+
+namespace SmartCon.FamilyManager.ViewModels;
+
+public sealed partial class FamilyManagerMainViewModel
+{
+    [RelayCommand]
+    private async Task OpenCategoryEditorAsync()
+    {
+        var editorVm = _viewModelFactory.CreateCategoryTreeEditorViewModel();
+        editorVm.Saved += () => _ = LoadTreeAsync();
+        await editorVm.InitializeAsync();
+        _dialogService.ShowCategoryTreeEditor(editorVm);
+    }
+
+    [RelayCommand]
+    private async Task EditMetadata()
+    {
+        if (SelectedItem is null) return;
+
+        var itemId = SelectedItem.Id;
+        var vm = _viewModelFactory.CreateMetadataEditViewModel(
+            SelectedItem.Id,
+            SelectedItem.Name,
+            SelectedItem.Description,
+            SelectedItem.CategoryId,
+            SelectedItem.CategoryName,
+            SelectedItem.Tags,
+            SelectedItem.ContentStatus);
+
+        var result = _dialogService.ShowMetadataEdit(vm);
+        if (result != true) return;
+
+        await LoadTreeAsync();
+        ExpandAndSelectItem(itemId);
+    }
+
+    [RelayCommand]
+    private async Task OpenProperties()
+    {
+        if (SelectedItem is null) return;
+
+        var itemId = SelectedItem.Id;
+        var updatedAt = SelectedItem.UpdatedAtUtc != default
+            ? SelectedItem.UpdatedAtUtc.ToLocalTime().ToString("yyyy-MM-dd HH:mm")
+            : null;
+
+        var vm = _viewModelFactory.CreatePropertiesViewModel(
+            SelectedItem.Id,
+            SelectedItem.Name,
+            SelectedItem.Description,
+            SelectedItem.CategoryId,
+            SelectedItem.CategoryName,
+            SelectedItem.Tags,
+            SelectedItem.ContentStatus,
+            SelectedItem.Manufacturer,
+            SelectedItem.VersionLabel,
+            null,
+            null,
+            updatedAt);
+
+        vm.InitializeCommand.Execute(null);
+        var result = _dialogService.ShowProperties(vm);
+        if (result != true) return;
+
+        await LoadTreeAsync();
+        ExpandAndSelectItem(itemId);
+    }
+
+    [RelayCommand]
+    private async Task UpdateFamilyAsync()
+    {
+        if (SelectedTreeNode is not FamilyLeafNodeViewModel leaf) return;
+
+        var title = LanguageManager.GetString(StringLocalization.Keys.FM_Update) ?? "Update";
+        var path = _dialogService.ShowOpenFileDialog(title);
+        if (path is null) return;
+
+        IsLoading = true;
+        try
+        {
+            var request = new FamilyUpdateRequest(leaf.CatalogItemId, path, CurrentRevitVersion);
+            var result = await _importService.UpdateFamilyAsync(request);
+
+            if (result.WasSkippedAsDuplicate)
+            {
+                StatusMessage = string.Format(
+                    LanguageManager.GetString(StringLocalization.Keys.FM_UpdateIdentical) ?? "File is identical to current version: {0}",
+                    result.FileName);
+            }
+            else if (result.Success)
+            {
+                StatusMessage = string.Format(
+                    LanguageManager.GetString(StringLocalization.Keys.FM_UpdateSuccess) ?? "Updated: {0} → {1}",
+                    result.FileName,
+                    result.VersionLabel);
+            }
+            else
+            {
+                StatusMessage = string.Format(
+                    LanguageManager.GetString(StringLocalization.Keys.FM_UpdateError) ?? "Update error: {0}",
+                    result.ErrorMessage);
+            }
+
+            await LoadTreeAsync();
+            ExpandAndSelectItem(leaf.CatalogItemId);
+        }
+        catch (Exception ex)
+        {
+            StatusMessage = string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_UpdateError) ?? "Update error: {0}",
+                ex.Message);
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+
+    [RelayCommand]
+    private async Task DeleteFamilyAsync()
+    {
+        if (SelectedItem is null) return;
+
+        var confirmed = _dialogService.ShowConfirmation(
+            LanguageManager.GetString(StringLocalization.Keys.FM_FamilyDeleteTitle) ?? "Delete Family",
+            string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_FamilyDeletePrompt) ?? "Delete \"{0}\"?",
+                SelectedItem.Name));
+
+        if (!confirmed) return;
+
+        IsLoading = true;
+        try
+        {
+            var success = await _writableProvider.DeleteItemAsync(SelectedItem.Id);
+            if (success)
+            {
+                StatusMessage = string.Format(
+                    LanguageManager.GetString(StringLocalization.Keys.FM_FamilyDeleted) ?? "Deleted: {0}",
+                    SelectedItem.Name);
+                await LoadTreeAsync();
+            }
+        }
+        catch (Exception ex)
+        {
+            StatusMessage = $"{LanguageManager.GetString(StringLocalization.Keys.FM_FamilyDeleteError) ?? "Error"}: {ex.Message}";
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+
+    public async Task MoveFamilyToCategoryAsync(string familyId, string? targetCategoryId)
+    {
+        IsLoading = true;
+        try
+        {
+            await _writableProvider.UpdateItemAsync(familyId, null, null, targetCategoryId, null, null);
+            await LoadTreeAsync();
+        }
+        catch (Exception ex)
+        {
+            StatusMessage = string.Format(
+                LanguageManager.GetString(StringLocalization.Keys.FM_ImportError) ?? "Error: {0}",
+                ex.Message);
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+
+    [RelayCommand(CanExecute = nameof(CanStartDrag))]
+    private void StartDrag(object? item)
+    {
+        // Drag permission gate. Behavior handles the actual DoDragDrop.
+    }
+
+    private bool CanStartDrag(object? item) => item is FamilyLeafNodeViewModel;
+
+    [RelayCommand(CanExecute = nameof(CanDropFamily))]
+    private async Task DropFamilyAsync(TreeViewDropInfo? info)
+    {
+        if (info is not { Payload: FamilyLeafNodeViewModel leaf, Target: CategoryNodeViewModel target })
+            return;
+
+        var categoryId = target.CategoryId == "__no_category__"
+            ? null
+            : target.CategoryId;
+
+        await MoveFamilyToCategoryAsync(leaf.CatalogItemId, categoryId);
+    }
+
+    private bool CanDropFamily(TreeViewDropInfo? info)
+    {
+        if (info is null) return false;
+        return info.Payload is FamilyLeafNodeViewModel
+            && info.Target is CategoryNodeViewModel;
+    }
+
+    private void ExpandAndSelectItem(string catalogItemId)
+    {
+        foreach (var root in TreeNodes)
+        {
+            if (ExpandToItem(root, catalogItemId))
+                return;
+        }
+    }
+
+    private bool ExpandToItem(CatalogTreeNodeViewModel node, string catalogItemId)
+    {
+        foreach (var child in node.Children)
+        {
+            if (child is FamilyLeafNodeViewModel leaf && leaf.CatalogItemId == catalogItemId)
+            {
+                node.IsExpanded = true;
+                leaf.IsSelected = true;
+                SelectedTreeNode = leaf;
+                return true;
+            }
+            if (ExpandToItem(child, catalogItemId))
+            {
+                node.IsExpanded = true;
+                return true;
+            }
+        }
+        return false;
+    }
+}
