+++ src/SmartCon.FamilyManager/ViewModels/FamilyPropertiesViewModel.cs
@@ -1,5 +1,5 @@
 using System.Collections.ObjectModel;
-using System.Diagnostics;
+using System.Linq;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using SmartCon.Core.Logging;
@@ -137,35 +137,6 @@ public sealed partial class FamilyPropertiesViewModel : ObservableObject, IObser
         finally
         {
             IsBusy = false;
-        }
-    }
-
-    private async Task LoadAssetsAsync(CancellationToken ct)
-    {
-        var assets = await _assetService.GetAssetsAsync(_catalogItemId, ct: ct);
-
-        ImageAssets = new ObservableCollection<FamilyAsset>(assets.Where(a => a.AssetType == FamilyAssetType.Image));
-        VideoAssets = new ObservableCollection<FamilyAsset>(assets.Where(a => a.AssetType == FamilyAssetType.Video));
-        DocumentAssets = new ObservableCollection<FamilyAsset>(assets.Where(a => a.AssetType == FamilyAssetType.Document));
-        LookupAssets = new ObservableCollection<FamilyAsset>(assets.Where(a => a.AssetType == FamilyAssetType.LookupTable));
-        SpreadsheetAssets = new ObservableCollection<FamilyAsset>(assets.Where(a => a.AssetType == FamilyAssetType.Spreadsheet));
-        Model3DAssets = new ObservableCollection<FamilyAsset>(assets.Where(a => a.AssetType == FamilyAssetType.Model3D));
-        OtherAssets = new ObservableCollection<FamilyAsset>(assets.Where(a => a.AssetType == FamilyAssetType.Other));
-
-        var primary = assets.FirstOrDefault(a => a.AssetType == FamilyAssetType.Image && a.IsPrimary);
-        if (primary is null)
-            primary = ImageAssets.FirstOrDefault();
-
-        if (primary is not null)
-        {
-            var path = await _assetService.ResolveAssetPathAsync(primary.Id, ct);
-            AvatarImagePath = path;
-            HasAvatar = path is not null;
-        }
-        else
-        {
-            AvatarImagePath = null;
-            HasAvatar = false;
         }
     }
 
@@ -336,121 +307,6 @@ public sealed partial class FamilyPropertiesViewModel : ObservableObject, IObser
 
     [RelayCommand]
     private void Cancel() => RequestClose?.Invoke(null);
-
-    [RelayCommand]
-    private async Task ChangeAvatar()
-    {
-        var path = _dialogService.ShowAssetOpenFileDialog(
-            LanguageManager.GetString(StringLocalization.Keys.FM_Props_SelectImage) ?? "Select image",
-            FamilyAssetType.Image);
-        if (path is null) return;
-
-        IsBusy = true;
-        try
-        {
-            var asset = await _assetService.AddAssetAsync(_catalogItemId, null, FamilyAssetType.Image, path, null);
-            await _assetService.SetPrimaryAssetAsync(asset.Id);
-            await LoadAssetsAsync(CancellationToken.None);
-        }
-        finally
-        {
-            IsBusy = false;
-        }
-    }
-
-    [RelayCommand]
-    private async Task RemoveAvatar()
-    {
-        if (!HasAvatar || ImageAssets.Count == 0) return;
-
-        var primary = ImageAssets.FirstOrDefault(a => a.IsPrimary) ?? ImageAssets.FirstOrDefault();
-        if (primary is null) return;
-
-        IsBusy = true;
-        try
-        {
-            await _assetService.DeleteAssetAsync(primary.Id);
-            await LoadAssetsAsync(CancellationToken.None);
-        }
-        finally
-        {
-            IsBusy = false;
-        }
-    }
-
-    [RelayCommand]
-    private async Task AddAsset(string assetTypeStr)
-    {
-        if (!Enum.TryParse<FamilyAssetType>(assetTypeStr, out var assetType)) return;
-
-        var path = _dialogService.ShowAssetOpenFileDialog(
-            LanguageManager.GetString(StringLocalization.Keys.FM_Props_AddFile) ?? "Select file",
-            assetType);
-        if (path is null) return;
-
-        IsBusy = true;
-        try
-        {
-            await _assetService.AddAssetAsync(_catalogItemId, null, assetType, path, null);
-            await LoadAssetsAsync(CancellationToken.None);
-        }
-        finally
-        {
-            IsBusy = false;
-        }
-    }
-
-    [RelayCommand]
-    private async Task DeleteAsset(FamilyAsset? asset)
-    {
-        if (asset is null) return;
-
-        IsBusy = true;
-        try
-        {
-            await _assetService.DeleteAssetAsync(asset.Id);
-            await LoadAssetsAsync(CancellationToken.None);
-        }
-        finally
-        {
-            IsBusy = false;
-        }
-    }
-
-    [RelayCommand]
-    private async Task OpenAsset(FamilyAsset? asset)
-    {
-        if (asset is null) return;
-
-        var path = await _assetService.ResolveAssetPathAsync(asset.Id);
-        if (path is null) return;
-
-        try
-        {
-            Process.Start(new ProcessStartInfo(path) { UseShellExecute = true });
-        }
-        catch (Exception ex)
-        {
-            SmartConLogger.Warn($"OpenAsset failed: {ex.Message}");
-        }
-    }
-
-    [RelayCommand]
-    private async Task SetAsPrimary(FamilyAsset? asset)
-    {
-        if (asset is null || asset.AssetType != FamilyAssetType.Image) return;
-
-        IsBusy = true;
-        try
-        {
-            await _assetService.SetPrimaryAssetAsync(asset.Id);
-            await LoadAssetsAsync(CancellationToken.None);
-        }
-        finally
-        {
-            IsBusy = false;
-        }
-    }
 }
 
 public sealed class FamilyTypeSelectorItem
